removed dead code + cleaning
