--- frontend/app/src/entities/ipam/ip-addresses/ipam-ip-address.tsx
@@ -1,16 +1,0 @@
-import { forwardRef } from "react";
-import { useParams } from "react-router";
-import IpAddressSummary from "./ip-address-summary";
-import IpamIPAddressesList from "./ipam-ip-address-list";
-
-const IpamIPAddresses = forwardRef((_, ref) => {
-  const { ip_address } = useParams();
-
-  if (ip_address) {
-    return <IpAddressSummary />;
-  }
-
-  return <IpamIPAddressesList ref={ref} />;
-});
-
-export default IpamIPAddresses;
