+++ frontend/app/src/entities/ipam/namespaces/ip-namespace-selector.tsx
