+++ frontend/app/src/entities/ipam/namespaces/ui/ip-namespace-provider.tsx
@@ -1,7 +1,7 @@
-import { constructPathForIpam } from "@/entities/ipam/common/utils";
 import { IPAM_QSP, IP_ADDRESS_GENERIC } from "@/entities/ipam/constants";
 import { IpNamespace } from "@/entities/ipam/namespaces/domain/get-ip-namespace-list";
 import { useGetIpNamespaceList } from "@/entities/ipam/namespaces/domain/get-ip-namespace-list.query";
+import { constructPathForIpam } from "@/entities/ipam/utils";
 import { NodeCore } from "@/entities/nodes/types";
 import { getSchema } from "@/entities/schema/domain/get-schema";
 import { isOfKind } from "@/entities/schema/utils/is-of-kind";
