--- frontend/app/src/entities/ipam/prefixes/ipam-prefixes-summary.tsx
@@ -1,16 +1,0 @@
-import { forwardRef } from "react";
-import { useParams } from "react-router";
-import IpamIPPrefixesSummaryDetails from "./ipam-prefixes-summary-details";
-import IpamIPPrefixesSummaryList from "./ipam-prefixes-summary-list";
-
-const IpamIPPrefixesSummary = forwardRef((_, ref) => {
-  const { prefix } = useParams();
-
-  if (prefix) {
-    return <IpamIPPrefixesSummaryDetails />;
-  }
-
-  return <IpamIPPrefixesSummaryList ref={ref} />;
-});
-
-export default IpamIPPrefixesSummary;
