--- frontend/app/src/entities/ipam/prefixes/ipam-summary-skeleton.tsx
@@ -1,23 +1,0 @@
-import { Skeleton } from "@/shared/components/skeleton";
-import { CardWithBorder } from "@/shared/components/ui/card";
-
-export const IpamSummarySkeleton = () => {
-  return (
-    <div className="flex items-start gap-2">
-      <CardWithBorder className="max-w-[479px] w-full">
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-        <Skeleton className="h-7 m-4" />
-      </CardWithBorder>
-    </div>
-  );
-};
