+++ frontend/app/src/entities/ipam/utils.ts
