+++ frontend/app/src/pages/ipam/ipam-ip-addresses-list-page.tsx
